refactor: 채팅방 생성 모달을 별도 컴포넌트로 분리 후 반영
