+++ src/pages/ChatPage/ChatPage.styled.ts
@@ -29,26 +29,3 @@ export const ChatRoomListWrapper = styled.div`
 		background: ${theme.color.border.secondary};
 	}
 `;
-
-export const ChatRoomModal = styled.div`
-	position: absolute;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	top: 50%;
-	left: 50%;
-	width: 400px;
-	height: 180px;
-	transform: translate(-50%, -50%);
-
-	border-radius: ${theme.units.radius.radius16};
-	z-index: ${theme.elevation.zIndex.MODAL};
-	gap: ${theme.units.spacing.space12};
-	background-color: ${theme.color.bg.primary};
-	box-shadow: ${theme.elevation.shadow.shadow16};
-
-	@media (max-width: 1440px) {
-		width: 240px;
-	}
-`;
